BackEndApp/requirements.txt
bcrypt==3.2.0
colorlog==5.0.1
cryptography==3.4.7
Django>=3.0.7
django-cors-headers==3.5.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
dlmml
dnspython==2.0.0
gunicorn==20.1.0
python-dotenv==0.15.0
pymongo==3.11.2
PyGithub==1.54.0.1
GitPython==3.1.18
# requirements-dev
pre-commit==2.13.0
black
autoflake
pytest
pytest-mock
mongomock
